+++ src/rra_tools/cli_tools/__init__.py
@@ -3,6 +3,8 @@ from rra_tools.cli_tools.importers import (
     import_module_from_info,
 )
 from rra_tools.cli_tools.options import (
+    process_choices,
+    with_choice,
     with_debugger,
     with_dry_run,
     with_input_directory,
@@ -24,4 +26,6 @@ __all__ = [
     "with_progress_bar",
     "with_queue",
     "with_verbose",
+    "process_choices",
+    "with_choice",
 ]