+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Remove `with_year` and add more generic `with_choice` option to `cli_tools` subpackage.
+- Add `process_choice` and `with_choice` to the `cli_tools` interface.
+
 
 ## [1.0.3] - 2024-05-13
 ### Added